Move Carousel to components folder
It is actually just a component, and not a stand-alone page.
